+++ src/api/devs/devs.types.ts
@@ -1,5 +1,26 @@
 import { LinkIconProps } from "../../lib/@me/LinkIcon/component";
-import { TagsEnum } from "./tags"
+
+export type TagsEnum = ([
+  "FUN", // memes, or other
+  "NEW_ELEMENTS",
+  "KEY_INPUTS",
+  "SHORTCUTS",
+  "PAGES",
+  "CALENDAR",
+  "NOTEBOOK",
+  "EXPLORER",
+  "WIDGETS",    // for additional modular UI components
+  "TASK_MANAGEMENT", // managing tasks in apps
+  "SEARCH", // search functionality
+  "USER_PROFILES", // user-related features
+  "NOTIFICATIONS", // for alerts or updates
+  "THEMES", // visual themes customization
+  "DASHBOARD", // central hub for accessing tools/pages
+  "NAVIGATION", // navigational improvements or shortcuts
+])[number]
+
+// type that has to have at least 1 element but can not have more than 3
+export type tags = Readonly<[TagsEnum, TagsEnum?, TagsEnum?]>;
 
 export interface Dev {
   /*
@@ -23,9 +44,6 @@ export interface Dev {
   description?: string;
   socials?: URL[];
 }
-
-// type that has to have at least 1 element but can not have more than 3
-export type tags = Readonly<[TagsEnum, TagsEnum?, TagsEnum?]>;
 
 // !import THE INTERFACE THAT EVERY EXTENSION MUST RETURN IN AN OBJECT
 export interface Plugin {
