--- src/api/devs/tags.ts
@@ -1,18 +1,0 @@
-export type TagsEnum = ([
-  "FUN", // memes, or other
-  "NEW_ELEMENTS",
-  "KEY_INPUTS",
-  "SHORTCUTS",
-  "PAGES",
-  "CALENDAR",
-  "NOTEBOOK",
-  "EXPLORER",
-  "WIDGETS",    // for additional modular UI components
-  "TASK_MANAGEMENT", // managing tasks in apps
-  "SEARCH", // search functionality
-  "USER_PROFILES", // user-related features
-  "NOTIFICATIONS", // for alerts or updates
-  "THEMES", // visual themes customization
-  "DASHBOARD", // central hub for accessing tools/pages
-  "NAVIGATION", // navigational improvements or shortcuts
-])[number]
